+++ atlas_splitter/app/cli.py
@@ -13,10 +13,6 @@ L = logging.getLogger(__name__)
 def cli():
     """The main CLI entry point"""
     logging.basicConfig(level=logging.INFO)
-    group = {
-        "layer-splitter": layer_splitter.app,
-    }
-    help_str = "The main CLI entry point."
-    app = click.Group("atlas_splitter", group, help=help_str)
+    app = layer_splitter.app
     app = click.version_option(VERSION)(app)
     app()