+++ atlas_splitter/exceptions.py
@@ -1,0 +1,9 @@
+"""
+Module dedicated to error handling.
+"""
+
+
+class AtlasSplitterError(Exception):
+    """
+    Exception raised by functions of atlas_splitter.
+    """