+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Changelog
 
 Version 0.1.0
 -------------
-- Initial commit: extract `layer_splitter` from `nse/atlas-building-tools` [`NSETM_1685_`]
+- Initial commit: extracts `region_splitter` and its tests from `nse/atlas-building-tools` [`NSETM_1685_`]
 
 
 .. _`NSETM-1685`: https://bbpteam.epfl.ch/project/issues/browse/NSETM-1685